show farmer in gg map by marker

+++ app/src/main/java/com/sun_asterisk/myfood/data/model/User.kt
@@ -31,5 +31,5 @@ data class User(
     val authorities: ArrayList<String>,
     @SerializedName("location")
     @Expose
-    var location: ArrayList<Long>?
+    var location: ArrayList<Float>?
 )
